# Avoid state machines