--- a/Fluid.Benchmarks/TagBenchmarks.cs
+++ b/Fluid.Benchmarks/TagBenchmarks.cs
@@ -4,7 +4,6 @@ using BenchmarkDotNet.Attributes;
 namespace Fluid.Benchmarks
 {
     [MemoryDiagnoser]
-    [ShortRunJob]
     public class TagBenchmarks
     {
         private static readonly FluidParser _parser = new();