--- a/Fluid/Ast/ElseIfStatement.cs
+++ b/Fluid/Ast/ElseIfStatement.cs
@@ -14,15 +14,50 @@ namespace Fluid.Ast
 
         public Expression Condition { get; }
 
-        public override async ValueTask<Completion> WriteToAsync(TextWriter writer, TextEncoder encoder, TemplateContext context)
+        public override ValueTask<Completion> WriteToAsync(TextWriter writer, TextEncoder encoder, TemplateContext context)
         {
             // Process statements until next block or end of statements
             for (var index = 0; index < _statements.Count; index++)
             {
                 context.IncrementSteps();
 
-                var completion = await _statements[index].WriteToAsync(writer, encoder, context);
+                var task = _statements[index].WriteToAsync(writer, encoder, context);
+                if (!task.IsCompletedSuccessfully)
+                {
+                    return Awaited(task, index + 1, writer, encoder, context);
+                }
 
+                var completion = task.Result;
+                if (completion != Completion.Normal)
+                {
+                    // Stop processing the block statements
+                    // We return the completion to flow it to the outer loop
+                    return new ValueTask<Completion>(completion);
+                }
+            }
+
+            return new ValueTask<Completion>(Completion.Normal);
+        }
+
+        private async ValueTask<Completion> Awaited(
+            ValueTask<Completion> task,
+            int startIndex,
+            TextWriter writer,
+            TextEncoder encoder,
+            TemplateContext context)
+        {
+            var completion = await task;
+            if (completion != Completion.Normal)
+            {
+                // Stop processing the block statements
+                // We return the completion to flow it to the outer loop
+                return completion;
+            }
+            // Process statements until next block or end of statements
+            for (var index = startIndex; index < _statements.Count; index++)
+            {
+                context.IncrementSteps();
+                completion = await _statements[index].WriteToAsync(writer, encoder, context);
                 if (completion != Completion.Normal)
                 {
                     // Stop processing the block statements