--- a/Fluid/Ast/IfStatement.cs
+++ b/Fluid/Ast/IfStatement.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.IO;
 using System.Text.Encodings.Web;
 using System.Threading.Tasks;