… Fluid/TrimmingFlags.cs